+++ CesiumGeospatial/test/TestBoundingRegionBuilder.cpp
@@ -4,6 +4,8 @@
 
 #include <catch2/catch.hpp>
 
+#include <cstring>
+
 using namespace CesiumGeospatial;
 using namespace CesiumUtility;
 
@@ -17,6 +19,10 @@ TEST_CASE("BoundingRegionBuilder") {
     CHECK(!rectangle.contains(Cartographic(0.0, 1.0, 0.0)));
     CHECK(!rectangle.contains(Cartographic(0.0, -1.0, 0.0)));
 
+    // The rectangle returned by toGlobeRectangle should be identical.
+    GlobeRectangle rectangle2 = builder.toGlobeRectangle();
+    CHECK(std::memcmp(&rectangle, &rectangle2, sizeof(GlobeRectangle)) == 0);
+
     builder.expandToIncludePosition(Cartographic(0.0, 0.0, 0.0));
     rectangle = builder.toRegion().getRectangle();
     CHECK(rectangle.contains(Cartographic(0.0, 0.0, 0.0)));
@@ -25,6 +31,9 @@ TEST_CASE("BoundingRegionBuilder") {
     CHECK(!rectangle.contains(Cartographic(0.0, 1.0, 0.0)));
     CHECK(!rectangle.contains(Cartographic(0.0, -1.0, 0.0)));
 
+    rectangle2 = builder.toGlobeRectangle();
+    CHECK(std::memcmp(&rectangle, &rectangle2, sizeof(GlobeRectangle)) == 0);
+
     builder.expandToIncludePosition(Cartographic(Math::OnePi, 1.0, 0.0));
     rectangle = builder.toRegion().getRectangle();
     CHECK(rectangle.contains(Cartographic(0.0, 0.0, 0.0)));
@@ -32,6 +41,9 @@ TEST_CASE("BoundingRegionBuilder") {
     CHECK(!rectangle.contains(Cartographic(-1.0, 0.0, 0.0)));
     CHECK(rectangle.contains(Cartographic(0.0, 1.0, 0.0)));
     CHECK(!rectangle.contains(Cartographic(0.0, -1.0, 0.0)));
+
+    rectangle2 = builder.toGlobeRectangle();
+    CHECK(std::memcmp(&rectangle, &rectangle2, sizeof(GlobeRectangle)) == 0);
 
     BoundingRegionBuilder simpleBuilder = builder;
     simpleBuilder.expandToIncludePosition(Cartographic(-1.0, 1.0, 0.0));
@@ -43,6 +55,9 @@ TEST_CASE("BoundingRegionBuilder") {
     CHECK(simple.contains(Cartographic(0.0, 1.0, 0.0)));
     CHECK(!simple.contains(Cartographic(0.0, -1.0, 0.0)));
 
+    rectangle2 = simpleBuilder.toGlobeRectangle();
+    CHECK(std::memcmp(&simple, &rectangle2, sizeof(GlobeRectangle)) == 0);
+
     BoundingRegionBuilder wrappedBuilder = builder;
     wrappedBuilder.expandToIncludePosition(Cartographic(-3.0, 1.0, 0.0));
     GlobeRectangle wrapped = wrappedBuilder.toRegion().getRectangle();
@@ -52,5 +67,8 @@ TEST_CASE("BoundingRegionBuilder") {
     CHECK(wrapped.contains(Cartographic(-3.0, 0.0, 0.0)));
     CHECK(wrapped.contains(Cartographic(0.0, 1.0, 0.0)));
     CHECK(!wrapped.contains(Cartographic(0.0, -1.0, 0.0)));
+
+    rectangle2 = wrappedBuilder.toGlobeRectangle();
+    CHECK(std::memcmp(&wrapped, &rectangle2, sizeof(GlobeRectangle)) == 0);
   }
 }